% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{as_numeric_factor}
\alias{as_numeric_factor}
\title{Numeric factor}
\usage{
as_numeric_factor(x)
}
\arguments{
\item{x}{Factor to coerce}
}
\value{
A vector as a numeric
}
\description{
Coerce a factor to a numeric vector
}
