% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B7.r
\docType{data}
\name{B7}
\alias{B7}
\title{Growth of Dutch boys}
\format{A data frame with 53 rows of 46 variables of a Behcet disease}
\description{
Height, weight, head circumference and puberty of 748 Dutch boys.
}
\details{
Random sample of 10\% from the cross-sectional data used to construct the
Dutch growth references 1997. Variables \code{gen} and \code{phb} are ordered
factors. \code{reg} is a factor.
}
\keyword{datasets}
