% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.result.r
\name{cox_result}
\alias{cox_result}
\title{Title}
\usage{
cox_result(obj, varname = NULL, test = "Likelihood")
}
\arguments{
\item{obj}{L'objet à présenter}

\item{varname}{Le nom de la variable}

\item{test}{Test à réaliser}
}
\value{
Une dataframe renvoyant la présentation d'un modèle de survie
}
\description{
Title
}
