% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.r
\name{logistic_multivariate}
\alias{logistic_multivariate}
\title{Multivariate Logistic Regression}
\usage{
logistic_multivariate(fit)
}
\arguments{
\item{fit}{Class glm. Multivariate model to format}
}
\value{
A dataframe of the multivariate parameters formatted
}
\description{
A function used to generate result table for multivariate logistic regression
model.
}
