% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.r
\name{make.language}
\alias{make.language}
\title{\code{make.language}}
\usage{
make.language(data, lang)
}
\arguments{
\item{data}{The table to be translated}

\item{lang}{The language to coerce to, default "en", "fr"}
}
\value{
The same table but translated
}
\description{
Exit language from description table
}
\examples{
make.language(AirPassengers, "fr")
}
