% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pval_format}
\alias{pval_format}
\title{p-value format}
\usage{
pval_format(pval)
}
\arguments{
\item{pval}{Numeric.}
}
\value{
"***" if < 0.001, "\*\*" if < 0.01, "\*" if < 0.05, "." if < 0.1
}
\description{
Format a p-value into R display system with stars
}
\examples{

pval_format(0.00025)

pval_format(0.20)

}
