% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.r
\name{step_lrcl_pval}
\alias{step_lrcl_pval}
\title{Backward stepwise selection with pvalue for logistic regression with
clustering}
\usage{
step_lrcl_pval(fitcl, cluster, threshold = 0.05, verbose = TRUE)
}
\arguments{
\item{fitcl}{Initial multivariate model}

\item{cluster}{Character. Name of the clustering variable of the model}

\item{threshold}{Numeric [0,1].}

\item{verbose}{Whether to display messages or not. Default TRUE}
}
\value{
A final multivariate model
}
\description{
Backward stepwise selection with pvalue for logistic regression with
clustering
}
