% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabkris.r
\name{TABKRIS}
\alias{TABKRIS}
\title{TABKRIS}
\usage{
TABKRIS(
  baz,
  vect.var,
  vect.quali,
  varint = NULL,
  valvarint = NULL,
  nomvarint = NULL,
  test = NULL,
  vecnoms = NULL,
  valeurs = NULL,
  vecrefs = NULL,
  varassoc = NULL,
  codassoc = NULL,
  pres = NULL,
  langue = "fr",
  digits = 2
)
}
\arguments{
\item{baz}{nom de table R}

\item{vect.var}{liste des noms de variables entre "" s?par?es par des virgules}

\item{vect.quali}{liste des types de variables s?par?es par des virgules
type=1 si     variable qualitative
type=0 sinon  (variable quantitative)}

\item{varint}{Variables d'intérêt}

\item{valvarint}{Liste des valeurs de la variable d'intérêt que l'on souhaite afficher dans le tableau}

\item{nomvarint}{Nom des valeurs de la variable d'intérêt}

\item{test}{Vecteurs des tests à passer dans la fonction TABKRIS}

\item{vecnoms}{Vecteur des noms des variables}

\item{valeurs}{Listes des valeurs que l'on souhaite afficher}

\item{vecrefs}{Modalité de référence pour les variables qualitatives}

\item{varassoc}{Variable imbriqué}

\item{codassoc}{Code de la variable imbriquée}

\item{pres}{Nombre de chiffres significatifs}

\item{langue}{Exit language, "en", "fr"}

\item{digits}{Nombre de décimales dans le tableau}
}
\value{
Rend une data frame contenant les informations de description
}
\description{
TABKRIS
}
