% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabkris2.r
\name{tabkris_2}
\alias{tabkris_2}
\title{Description of data}
\usage{
tabkris_2(
  data,
  names = NULL,
  varint = NULL,
  lang = "en",
  method = NULL,
  test = FALSE,
  default_method = c("cont", "bino", "cate", "ordo"),
  default_test = c("stud", "chisq", "chisq", "chisq"),
  explicit_na = FALSE,
  digits = NULL,
  return_table = TRUE
)
}
\arguments{
\item{data}{Dataframe to describe}

\item{names}{Vectors of variables to display in the final table, length of
data columns}

\item{varint}{Variable to stratify on, factor only}

\item{lang}{Language to display, default "en", "fr"}

\item{method}{Vectors of variables to customize the methods used for
description, length of data columns}

\item{test}{Vectors of variables to customize the tests used for statistics,
length of data columns}

\item{default_method}{Default method to compute the table for each variable}

\item{default_test}{Default test to apply for each variable type}

\item{explicit_na}{Whether to account for NA in description}

\item{digits}{Number of significant number to display, default system option}

\item{return_table}{Whether to return a dataframe or an object to customize
option easily, default TRUE}
}
\value{
A dataframe or an object with all arguments to customize function
  call
}
\description{
\code{tabkris_2} returns a dataframe with descriptive statistics for each
variable
}
\details{
The \code{tabkris_2} function is a function to describe a set of
  data. Main purpose is to create a typical table one in biomedical
  litterature, either a patient characteristic table or population
  characteristic table. \code{names} is a vector to name the variable of
  data. Default will use the colnames of data. \code{varint} is a variable to
  stratify the analysis. It must be included in the initial dataset. It will
  not be displayed in the final table if chosen as the stratifying variable
  \code{lang} is useful to choose the language for the final display. The
  default is english. French is also supported. \code{explicit_na}.
}
