% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{as_numeric_factor}
\alias{as_numeric_factor}
\title{Numeric factor}
\usage{
as_numeric_factor(x)
}
\arguments{
\item{x}{Factor to coerce}
}
\value{
A vector as a numeric
}
\description{
Coerce a factor to a numeric vector
}
\examples{

vec <- as.factor(c(8:10))

as.numeric(vec) # Return a false value
as_numeric_factor # Return the value of numeric vector

}
