% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_univariate.r
\name{logistic_univariate}
\alias{logistic_univariate}
\title{Univariate Logistic}
\usage{
logistic_univariate(data, y, x, twobytwo = TRUE, formula = "(y ~ x)")
}
\arguments{
\item{data}{A dataframe including all the variable needed in all the models}

\item{y}{A vector of the names of response variable}

\item{x}{A vector of the names of predictor variable}

\item{twobytwo}{Either to display the two by two table for each. Default is \code{TRUE}}

\item{formula}{Formula for logistic regression to customize. Default is \code{(x ~ y)}}
}
\value{
If \code{lenght(y) > 1}, the returned value is a list including all
  table result for each response variable. If \code{lenght(y) > 1}, the
  returned value is a dataframe object, the table result.
}
\description{
A function used to generate multiple result table for univariate logistic
regression model with \code{y ~ x}. For each specified \code{y}, a result
table is computed, including every \code{x} variable.
}
