% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{select_noms}
\alias{select_noms}
\title{Name selection}
\usage{
select_noms(fichier)
}
\arguments{
\item{fichier}{Le fichier a passer}
}
\value{
Permet de sélectionner les noms dans un vecteur
}
\description{
Name selection
}
