% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabkris2.r
\name{tabkris_2}
\alias{tabkris_2}
\title{Description of data}
\usage{
tabkris_2(
  data,
  names = NULL,
  varint = NULL,
  lang = "en",
  method = NULL,
  test = FALSE,
  pres_quant = c("mean"),
  pres_quali = c("n", "per"),
  default_method = c("cont", "bino", "cate", "ordo"),
  default_test = c("stud", "chisq", "chisq", "chisq"),
  explicit_na = FALSE,
  digits = 2,
  return_table = TRUE,
  auto_detect = FALSE
)
}
\arguments{
\item{data}{Dataframe to describe}

\item{names}{Vectors of variables to display in the final table, length of
\code{ncol(data)}}

\item{varint}{Variable to stratify on, factor only}

\item{lang}{Language to display, default \code{"en"}, \code{"fr"}}

\item{method}{Vectors of variables to customize the methods used for
description, length of data columns}

\item{test}{Either a logical indicating statistical tests execution or a
vectors of variables to customize the tests, length of data columns.
Default \code{FALSE}}

\item{pres_quant}{Descriptive statistics for quantitative variables. Possible
values are \code{"mean"} for mean, SD, \code{"med"} for median, IQR,
\code{"range"} for range}

\item{pres_quali}{Descriptive statistics for qualitative variables. Possible
values are \code{"n"} for number, \code{"total"} to add "/ total" and
\code{"per"} for percentages}

\item{default_method}{Default method to compute the table for each variable.
Default \code{default_method = c("cont", "bino", "cate", "ordo")}}

\item{default_test}{Default test to apply for each variable type. Default
\code{c("stud", "chisq", "chisq", "chisq")}. Available \code{"stud",
"wilcox", "kruskal", "chisq", "fish"}}

\item{explicit_na}{Whether to display NA in description, Default \code{FALSE}}

\item{digits}{Number of significant number to display, default \code{2}}

\item{return_table}{Whether to return a dataframe or an object to customize
option easily, default \code{TRUE}}

\item{auto_detect}{Whether to automatically detect variable type,
transforming to factors numeric variable with moderate levels (< 10),
default \code{FALSE}}
}
\value{
A dataframe or an object with all arguments to customize function
  call
}
\description{
\code{tabkris_2} returns a dataframe with descriptive statistics for each
variable
}
\details{
The \code{tabkris_2} function is a function to describe a set of
  data. Main purpose is to create a typical table one in biomedical
  litterature, either a patient characteristic table or population
  characteristic table. \code{names} is a vector to name the variable of
  data. Default will use the colnames of data. \code{varint} is a variable to
  stratify the analysis. It must be included in the initial dataset. It will
  not be displayed in the final table if chosen as the stratifying variable
  \code{lang} is useful to choose the language for the final display. The
  default is english. French is also supported. \code{explicit_na}.
}
